requirements.txt
google-api-python-client>=1.7.4
oauth2client>=4.1.2
pandas>=0.23.0
pyqt>=5.9.2
